Save reset value to EEPROM if browned out
Previously, if reset was pressed while browned out, the in-RAM
instance of _lastPixel was reset properly, but when the system
came out of brown out, the restore from EEPROM would happen,
and the reset value over-written with the last saved value.

I modified VoltMode::reset to check if a brown out was in
progress, and save the zero value if that is the case.  Then
when the system comes out of brown out, it will get that
reset value.

The condition to check if brownout was in progress was added
to prevent unnecessary writes to EEPROM when reset is pressed
when not browned out.